Re-implement legacy SynthIcon class
SeaGlassLookAndFeel uses a couple of legacy classes in the
now-deprecated (and -removed) `sun.swing` namespace.

Most of these components were re-implemented/moved to the default
`javax.swing` namespace in newer Java releases, but, because those old
Swing classes were no true part of the Java standard library, breaking
changes were introduced in the process, and some classes have not been
migrated.

The sun.wing.plaf.synth.SynthIcon class has been moved, but was
converted from an abstract class to an interface, and three static
methods were not migrated, meaning that the compatibility issues can
not be resolved by just replacing the legacy class with the migrated
version.
To solve this, the legacy SynthIcon class has been re-implemented (so
no changes have to be made to the main codebase) by just implementing
the new SynthIcon interface and re-implementing the missing static
methods from scratch.