Add sun.awt.AppContext implementation
Parts of the SeaGlassLookAndFeel codebase rely on access to the
internal AWT AppContext to cache data. However, starting with Java 9
there is no way to easily access the AppContext anymore, and, unlike
other parts of the inofficial AWT/Swing classes, its functionality
can not easily be replicated.

In order to at least make the project compile again,
a temporary barebones realisation of the functionality required by
the codebase has been implemented.