Add KeybindingsParser class and tests
SeaGlassLookAndFeel uses the now-inaccessible classes
WindowsKeybindings and GTKKeybindings to set the keybindings for the
Windows and GTK platforms, respectively (macOS keybindings are set via
an internal MacKeybindings class).

The original implementations of these classes create/install the
keybindings manually, which is not only tedious and inelegant, but also
hard to refactor and edit in general.
For the re-implementation, the keybindings will be stored as a resource
using a new internal file format (.kb), and will be loaded/created at
runtime. The parsing of these keybinding files is performed by the
new KeybindingsParser class. A couple of tests for this new class have
also been implemented as KeybindingsParserTest.