Re-implement WindowsKeybindings and add tests
Re-implement the WindowsKeybindings utility class using the new
KeybindingsLoader. The re-implementation simply loads the keybindings
from the 'Windows.kb' resource.

Its test simply ensures the new implementation is consistent with the
old one (produces the same results).

To compare different keybindings, a custom utility function in
AssertUtils.kt, assertKeybindingsEquals() had to be implemented because
of problems with LazyInputMap's implementation of the standard equals()
method.